requirements.txt
requests==2.32.3
Django==5.0.2
dj-rest-auth==4.0.0
django-allauth==0.52.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-filter==23.2
django-storages==1.14.3
dj-database-url==2.2.0
psycopg[c]==3.1.18
python-dotenv==1.0.1
langchain==0.2.15
langchain_community==0.2.14
langchain_core==0.2.36
langchain_openai==0.1.23
langchain_text_splitters==0.2.2
beautifulsoup4==4.12.3
gunicorn==20.1.0
azure-core==1.30.2
azure-storage-blob==12.20.0
pillow==9.5.0